growbuddiesproject/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "growbuddies"
version = "2022.12.7"
readme = "README.md"
description = "Buddies to help maximize the growth of plants in your indoor garden."
dependencies = [
    "influxdb ~=5.3.1",
    "paho_mqtt ~=1.6.1",
]

[project.urls]
"Homepage" = "https://github.com/solarslurpi/GrowBuddies"

[project.scripts]
store-readings = "growbuddies.__main__:main"

[tool.black]
line-length = 119

